feat: Add configurable check dictionary
If the build machine has no connection to the public internet, the
`clang_tidy_visualizer` fails. Because the `find_checks_dict` is unable
to generate the `checks_dict`.

This problem can be solved if an offline or internal mirror can be set
as the source of the dictionary.

From now, an optional `--checks_dict_url` is added to the arguments.
The argument is set the url of the dictionary. In case of missing url,
the `clang_tidy_visualizer` works the original way.

The documentation links are left intact.